requirements.txt
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
pytesseract==0.3.5
tensorflow==2.2.0

